+++ src/collective/z3cform/colorpicker/demo/browser.py
@@ -1,19 +1,18 @@
 from plone.z3cform.layout import wrap_form
 from z3c.form import form, field, button
-from zope import schema
 from zope.interface import Interface
-from collective.z3cform.colorpicker import ColorpickerFieldWidget
-from collective.z3cform.colorpicker import ColorpickerAlphaFieldWidget
+from collective.z3cform.colorpicker import Color
+from collective.z3cform.colorpicker import ColorAlpha
 
 
 class IColorForm(Interface):
-    color = schema.TextLine(
+    color = Color(
         title=u"Color",
         description=u"",
         required=False,
         default=u"#6298c9")
 
-    alphacolor = schema.TextLine(
+    alphacolor = ColorAlpha(
         title=u"Color with alpha layer support",
         description=u"",
         required=False,
@@ -24,8 +23,6 @@ class ColorDemoForm(form.Form):
     """Example color picker form"""
 
     fields = field.Fields(IColorForm)
-    fields['color'].widgetFactory = ColorpickerFieldWidget
-    fields['alphacolor'].widgetFactory = ColorpickerAlphaFieldWidget
     ignoreContext = True
 
     label = u"Colopicker Demo"